reafactor: change money BRL func folder and logic in form
